2008-06-01 23:56 UTC+0100 Viktor Szakats (harbour.01 syenar hu)    * include/hbvm.h    * source/vm/hvm.c    * source/vm/extrap.c      ! Fixed OS/2 problem reported by David.        (BTW, do we need really need to remove the exception        handler before exiting? If yes, shouldn't we do the        same under Win32?)
   * source/common/hbstr.c
     ! Silenced one 'loss of significant digit' BCC warning.

+++ harbour/source/common/hbstr.c
@@ -554,7 +554,7 @@ static BOOL hb_str2number( BOOL fPCode, const char* szNum, ULONG ulLen, HB_LONG 
       int iLimC;
 
       lLimV = HB_LONG_MAX / 10;
-      iLimC = HB_LONG_MAX % 10;
+      iLimC = ( int ) ( HB_LONG_MAX % 10 );
 
       iWidth = ulPos;
